Bild Eigen for arm64

+++ IOS/Sample/IdoorSample/Library/include/IndoorLibrary/MeasurementEvent.h
@@ -19,7 +19,7 @@ typedef NS_ENUM(NSInteger, MeasurementType)
 };
 
 @interface MeasurementEvent : NSObject
-@property (nonatomic, assign) NSInteger timestamp;
+@property (nonatomic, assign) CGFloat timestamp;
 @property (nonatomic, assign) MeasurementType type;
 @property (nonatomic, strong) NSArray *values;
 @property (nonatomic, assign) CGFloat latitude;
